# ENH: linalg.exp: Improve performance for small matrices