--- a/scipy/linalg/_matfuncs_expm.pyx.in
+++ b/scipy/linalg/_matfuncs_expm.pyx.in
@@ -146,10 +146,11 @@ def pick_pade_structure(cnp.ndarray[ndim=3, dtype=numpy_lapack_t] Am):
     # m = 3
     # -------
     # 1-norm of A ** 7
-    for i in range(4):
+    for i in range(3):
         np.matmul(absA, work_arr, out=work_arr)
     absA_pow4 = work_arr.copy()
-    np.matmul(absA_pow4, work_arr, out=work_arr)
+    for i in range(3):
+        np.matmul(absA, work_arr, out=work_arr)
     temp = work_arr.max()
 
     lm = max(<int>ceil(log2(temp/normA/coeff[0])/6), 0)