--- a/scipy/linalg/_matfuncs_expm.pyx.in
+++ b/scipy/linalg/_matfuncs_expm.pyx.in
@@ -150,7 +150,6 @@ def pick_pade_structure(cnp.ndarray[ndim=3, dtype=numpy_lapack_t] Am):
     np.matmul(absA, work_arr, out=work_arr)
     absA_pow4 = Am[1, :, :] # use the slice Am[1] as a working buffer
     np.matmul(absA, work_arr, absA_pow4)
-    del absA
     np.matmul(absA_pow4, work_arr, out=work_arr)
     temp = work_arr.max()