For small matrices the method `pick_pade_structure` can use a significant amount of time in `scipy.linalg.expm`. Part of the calculation inside `pick_pade_structure` is the calculation of the `absA**k` for various values of `k` (7, 11, 15, 19, 27) . Instead of calculating this power by repeated multiplication with `absA`, we calculate `absA**4` once and use that for calculation of higher order values. We also replace `np.max(work_arr)` with `work_arr.max()` (for small arrays the latter has significantly less overhead).

Benchmark:
```
import time
import scipy
from scipy.linalg import expm
import numpy as np
print(scipy, scipy.__version__)

for size in [2, 3, 4]:
    a=np.random.rand(5000, size, size)
    
    t0=time.perf_counter()
    
    for ii in range(len(a)):
        A = a[ii,:,:]
        expm(A)
        
    dt=time.perf_counter()-t0
    print(f'{size=} {dt:.3f} [s]')
```
Results on main:
```
<module 'scipy' from '/home/eendebakpt/sc/lib/python3.10/site-packages/scipy/__init__.py'> 1.15.0.dev0+git20240903.cd25eb1
size=2 0.567 [s]
size=3 0.649 [s]
size=4 0.689 [s]
```
Results on this PR:
```
<module 'scipy' from '/home/eendebakpt/sc/lib/python3.10/site-packages/scipy/__init__.py'> 1.15.0.dev0+git20240903.efde44b
size=2 0.401 [s]
size=3 0.430 [s]
size=4 0.447 [s]
```

<!-- 
Thanks for contributing a pull request! Please ensure that
your PR satisfies the checklist before submitting:
https://scipy.github.io/devdocs/dev/contributor/development_workflow.html#checklist-before-submitting-a-pr

Also, please name and describe your PR as you would write a
commit message:
https://scipy.github.io/devdocs/dev/contributor/development_workflow.html#writing-the-commit-message.
However, please only include an issue number in the description, not the title,
and please ensure that any code names containing underscores are enclosed in backticks.

Depending on your changes, you can skip CI operations and save time and energy: 
https://scipy.github.io/devdocs/dev/contributor/continuous_integration.html#skipping

Note that we are a team of volunteers; we appreciate your
patience during the review process.

Again, thanks for contributing!
-->

#### Reference issue

None

#### What does this implement/fix?

The PR improves performance of `expm`