@eendebakpt 2024-09-03
ASV results show impact on larger matrices as well.

Main:
```
[ 0.00%] · For scipy commit cd25eb1e <main>:
[ 0.00%] ·· Benchmarking virtualenv-py3.10-Cython-meson-python-numpy-pybind11-pytest-pythran
[25.00%] ··· Running (sparse_linalg_expm.Expm.time_expm--)..
[75.00%] ··· sparse_linalg_expm.Expm.time_expm                                                                                                                      ok
[75.00%] ··· ===== ============ ============
             --              format         
             ----- -------------------------
               n      sparse       dense    
             ===== ============ ============
               30   12.2±0.2ms   1.21±0.2ms 
              100    42.4±3ms    9.87±20ms  
              300    358±3ms      227±5ms   
             ===== ============ ============

[100.00%] ··· sparse_linalg_expm.ExpmMultiply.time_expm_multiply  
```
PR
```
[ 0.00%] · For scipy commit 8c367d48 <exmp>:
[ 0.00%] ·· Benchmarking virtualenv-py3.10-Cython-meson-python-numpy-pybind11-pytest-pythran
[25.00%] ··· Running (sparse_linalg_expm.Expm.time_expm--)..
[75.00%] ··· sparse_linalg_expm.Expm.time_expm                                                                                                                      ok
[75.00%] ··· ===== ========== ===========
             --            format        
             ----- ----------------------
               n     sparse      dense   
             ===== ========== ===========
               30   10.9±2ms   535±100μs 
              100   41.9±4ms    12.8±1ms 
              300   307±10ms   219±0.2ms 
             ===== ========== ===========

[100.00%] ··· sparse_linalg_expm.ExpmMultiply.time_expm_multiply 
```