@ilayn 2024-09-04
That's a very nice observation indeed. 

After 3rd power, you don't need absA either since increment is of 4 after that. So you hold 3rd power and 4th and get 7 in one go eliminating two more matmuls which is the actual time killer. Then you can eliminate the extra array (i think) which is an issue for large arrays of 1000s. Matmul already makes a temp copy internally.

At that stage later slices of Am is empty so you can dump it temporarily there which is already allocated.